@@ -38,7 +38,7 @@ const filePathFor = (stateNavigator, key, params) => {
   const route = stateNavigator.getNavigationLink(key, params);
   if (!route) {
     throw new Error(
-      `The route could not be matched for key: ${key}, params: ${JSON.stringify(params)}`
+      `The route could not be matched for key: ${key}, params: ${JSON.stringify(params)}`,
     );
   }
   return { link: route, filePath: routeFilePath(route.substring(1)) };
