@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import Header from '../components/header';
 import Footer from '../components/footer';
 import styles from './style.css';
