@@ -5,7 +5,7 @@ import QAndASlice from '.';
 describe('site/about-us/q-and-a-slice', () => {
   it('should render nothing for empty array', () => {
     const qAndASlice = render(<QAndASlice qAndAs={[]} />);
-    expect(qAndASlice.html()).to.equal('');
+    expect(qAndASlice.html()).to.equal(null);
   });
 
   it('should render heading 2', () => {
