@@ -1,11 +1,23 @@
+// @flow
 import React, { Component } from 'react';
 import classnames from 'classnames/bind';
 import styles from '../style.css';
 
 const cx = classnames.bind(styles);
-const { func, string, bool } = React.PropTypes;
 
-const AfterSignUpForm = ({ errorMessage, handleInputChange, submitting, handleSubmit }) => (
+type AfterSignUpFormProps = {
+  errorMessage: string,
+  handleInputChange: Function,
+  submitting: boolean,
+  handleSubmit: Function,
+};
+
+const AfterSignUpForm = ({
+  errorMessage,
+  handleInputChange,
+  submitting,
+  handleSubmit,
+}: AfterSignUpFormProps) => (
   <div className={styles.newsletterInnerAfter}>
     <h2 className={styles.subTitle}>
       Help us make sure your BadgerNews is relevant by telling us a bit more about yourself
@@ -83,21 +95,21 @@ const AfterSignUpForm = ({ errorMessage, handleInputChange, submitting, handleSu
   </div>
 );
 
-AfterSignUpForm.propTypes = {
-  errorMessage: string.isRequired,
-  handleInputChange: func.isRequired,
-  submitting: bool.isRequired,
-  handleSubmit: func.isRequired,
+type AfterSignupProps = {
+  onSubmit: Function,
+  updatedFormSubmitted: boolean,
+  errorMessage: string,
 };
 
-export default class AfterSignup extends Component {
-  static propTypes = {
-    onSubmit: func.isRequired,
-    updatedFormSubmitted: bool.isRequired,
-    errorMessage: string.isRequired,
-  };
+type AfterSignupState = {
+  name: string,
+  company: string,
+  role: string,
+  submitting: boolean,
+};
 
-  constructor(props) {
+export default class AfterSignup extends Component<AfterSignupProps, AfterSignupState> {
+  constructor(props: AfterSignupProps) {
     super(props);
     this.state = {
       name: '',
@@ -117,7 +129,7 @@ export default class AfterSignup extends Component {
     });
   }
 
-  handleInputChange = event => {
+  handleInputChange = (event: SyntheticInputEvent<HTMLInputElement>) => {
     const newState = {};
     newState[event.target.name] = event.target.value;
     this.setState(newState);
@@ -132,8 +144,11 @@ export default class AfterSignup extends Component {
 
   // fixes bug in safari where the component height wouldn't update
   triggerReflow() {
+    // $FlowIgnore
     this.element.style.display = 'none';
+    // $FlowIgnore
     this.element.offsetHeight; // eslint-disable-line no-unused-expressions
+    // $FlowIgnore
     this.element.style.display = '';
   }
 
@@ -142,6 +157,7 @@ export default class AfterSignup extends Component {
       <section
         className={cx('newsletter', 'submitted')}
         ref={c => {
+          // $FlowIgnore
           this.element = c;
         }}
       >
