@@ -1,3 +1,4 @@
+// @flow
 import React, { Component } from 'react';
 import classnames from 'classnames/bind';
 import styles from '../style.css';
@@ -7,8 +8,23 @@ import groups from '../../../../../services/mailchimp/config';
 
 const cx = classnames.bind(styles);
 
-class BeforeSignUp extends Component {
-  constructor(props) {
+type BeforeSignUpProps = {
+  onSubmit: Function,
+  errorMessage: string,
+};
+
+type InterestsState = {
+  [groups.techPageSignup]: boolean,
+};
+
+type BeforeSignUpState = {
+  email_address: string,
+  submitting: boolean,
+  interests: InterestsState,
+};
+
+class BeforeSignUp extends Component<BeforeSignUpProps, BeforeSignUpState> {
+  constructor(props: BeforeSignUpProps) {
     super(props);
     this.state = {
       email_address: '', // eslint-disable-line camelcase
@@ -25,7 +41,7 @@ class BeforeSignUp extends Component {
     });
   }
 
-  handleInputChange = event => {
+  handleInputChange = (event: SyntheticInputEvent<HTMLInputElement>) => {
     this.setState({
       [event.target.name]: event.target.value,
     });
@@ -99,11 +115,5 @@ class BeforeSignUp extends Component {
     );
   }
 }
-
-const { func, string } = React.PropTypes;
-BeforeSignUp.propTypes = {
-  onSubmit: func.isRequired,
-  errorMessage: string.isRequired,
-};
 
 export default BeforeSignUp;
