@@ -132,7 +132,7 @@ const handleContactUsHash = stateNavigator => {
 export default () => {
   const stateNavigator = new StateNavigator(
     routes(),
-    new HTML5HistoryManager((process.env.URL_BASENAME || '').slice(0, -1))
+    new HTML5HistoryManager((process.env.URL_BASENAME || '').slice(0, -1)),
   );
   handleContactUsHash(stateNavigator);
   stateNavigator.onNavigate((oldState, state) => {
